modifs

+++ web/admin/messagerie/suppr_message.php
@@ -8,12 +8,17 @@ require_once('../inc/droits.inc.php');
 
 <?php
 
-$id=$_GET['id'];
-$db = new DB_connection();
+if(isset($_GET['id']))
+{
+	$id=$_GET['id'];
+	$db = new DB_connection();
 
-$req="DELETE from Message where id_message='".$id."'";
+	$req="DELETE from Message where id_message='".$id."'";
 
-$db->DB_query($req);
+	$db->DB_query($req);
+
+	$db->DB_done();
+}
 	
 $url="messagerie.php";
 header("Refresh:0;url=$url");
